moved front end token authentication utils/auth from server to client's, added/additional dir and sample files reducers, actions, UserContext under context/
