+++ client/src/context/actions.js
@@ -1,0 +1,3 @@
+// change of states for login/logout being exported out into src/utils/auth.js
+export const LOGIN = "LOGIN";
+export const LOGOUT = "LOGOUT";
